% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{mistral_api_key}
\alias{mistral_api_key}
\title{Retrieve the Mistral API key}
\usage{
mistral_api_key()
}
\value{
The stored API key as a string.
}
\description{
This function retrieves the Mistral API key stored in the R options.
}
