% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload.R
\name{mistral_get_file_url}
\alias{mistral_get_file_url}
\title{Get a signed URL for file access}
\usage{
mistral_get_file_url(file_id, expiry = 24, api_key = mistral_api_key())
}
\arguments{
\item{file_id}{ID of the file to get URL for}

\item{expiry}{Expiry time in hours, default is 24}

\item{api_key}{Your MistralAI API key}
}
\value{
The signed URL information as a list
}
\description{
Get a signed URL for file access
}
