% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{mistral_ocr_save_md}
\alias{mistral_ocr_save_md}
\title{Save OCR results as a Markdown file}
\usage{
mistral_ocr_save_md(
  ocr,
  file,
  by_page = FALSE,
  save_images = TRUE,
  overwrite = FALSE,
  img_dir = dirname(file)
)
}
\arguments{
\item{ocr}{A list containing OCR results, including extracted text and images.}

\item{file}{The file path where the Markdown content should be saved.}

\item{by_page}{Logical. If `TRUE`, saves each page as a separate file. Defaults to `FALSE`.}

\item{save_images}{Logical. If `TRUE`, also saves extracted images. Defaults to `TRUE`.}

\item{overwrite}{Logical. If `TRUE`, overwrites existing files. Defaults to `FALSE`.}

\item{img_dir}{Directory where images should be saved. Defaults to the directory of `file`.}
}
\description{
This function saves the extracted OCR text in Markdown format, either as a single file or split by pages.
}
