% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_mistral_api_key}
\alias{set_mistral_api_key}
\title{Set the Mistral API key}
\usage{
set_mistral_api_key(key = NULL, file = NULL)
}
\arguments{
\item{key}{Optional. The API key as a string.}

\item{file}{Optional. A file path containing the API key.}
}
\description{
This function sets the Mistral API key, either from a provided string or by reading from a file.
}
